+++ docs/about/history.rst
@@ -34,6 +34,7 @@ Fixed
    -  Limit minimum phase period of flow technique applied to waves to fix objectionable phasing issues in flowing water like rivers.
    -  Fixed some components breaking in edit mode after entering/exiting prefab mode.
    -  Fixed *Build Processor* deprecated/obsolete warnings.
+   -  Fixed spurious "headless/batch mode" error during builds.
 
 
 4.15.2